refactor: Refactored PublicationForm Files
Ordered publication form files correctly for testing
