+++ src/components/PublicationRequestFormSections/PublicationForm/PublicationBook/index.js
@@ -1,0 +1,1 @@
+export { default } from './PublicationBook';
