+++ src/components/PublicationRequestFormSections/PublicationForm/PublicationJournal/index.js
@@ -1,0 +1,1 @@
+export { default } from './PublicationJournal';
