+++ src/components/PublicationRequestFormSections/PublicationForm/index.js
@@ -1,1 +1,3 @@
-export { default } from './PublicationForm';
+export { default as PublicationForm } from './PublicationForm';
+export { default as PublicationJournal } from './PublicationJournal';
+export { default as PublicationBook } from './PublicationBook';
