+++ src/components/PublicationRequestFormSections/index.js
@@ -1,5 +1,5 @@
 export { default as FundingForm } from './FundingForm';
-export { default as PublicationForm } from './PublicationForm';
+export { default as PublicationForm } from './PublicationForm/PublicationForm';
 export { default as PublicationStatusForm } from './PublicationStatusForm/PublicationStatusForm';
 export { default as RequestInfoForm } from './RequestInfoForm';
 export { default as RequestContactForm } from './RequestContactForm';
